feat: добавил разграничение пользователя

+++ back/app/Http/Controllers/WorkspaceController.php
@@ -23,7 +23,7 @@ class WorkspaceController extends Controller
 
     public function show(Workspace $workspace)
     {
-        $userRole = $this->workspaceService->getUserRoleInWorkspace($workspace, auth()->user());
+        $userRole = $this->workspaceService->getUserRoleInWorkspace($workspace, auth()->id());
 
         return response()->json([
             'workspace' => $workspace,
@@ -33,15 +33,20 @@ class WorkspaceController extends Controller
 
     public function store(Request $request)
     {
-        $title = $request->title ?? 'new Workspace';
+        $validate = $request->validate([
+            'title' => 'required|string|max:255'
+        ]);
+
+        $title = $validate['title'] ?? 'new Workspace';
 
         $workspace = $this->workspaceService->createWorkspace($title);
 
-        return response()->json(['massage' => 'добавление успешно'], 201);
+        return response()->json(['message' => __('messages.add_success')], 201);
     }
 
     public function update(Request $request, Workspace $workspace)
     {
+        $this->authorize('manage', $workspace);
         $validate = $request->validate([
             'title' => 'required|string|max:255',
             'description' => 'nullable|string',
@@ -54,35 +59,52 @@ class WorkspaceController extends Controller
 
     public function destroy(Workspace $workspace)
     {
+        if (!$workspace) {
+            return response()->json(['message' => 'Workspace not found'], 404);
+        }
+
         $this->workspaceService->deleteWorkspace($workspace);
 
-        return response()->json(['message' => 'удаление успешно'], 202);
+        return response()->json(['message' => __('messages.delete_success')], 202);
     }
 
     public function workspaceUsers(Workspace $workspace)
     {
-        return response()->json($workspace->users);
+        $workspaceWithUsers = $workspace->load('users');
+        $sanitizedUsers = $workspaceWithUsers->users->map(function ($user) {
+            return [
+                'id' => $user->id,
+                'name' => $user->name,
+                'email' => $user->email,
+            ];
+        });
+        return response()->json($sanitizedUsers);
     }
 
     public function manageUserInWorkspace(Request $request, Workspace $workspace)
     {
+        $validate = $request->validate([
+            'user_id' => 'required|integer',
+            'role' => 'string|in:owner,admin,project_manager,executor'
+        ]);
+
         $currentUser = $request->user();
 
-        $currentUserRole = $this->workspaceService->getUserRoleInWorkspace($workspace, $currentUser);
+        $currentUserRole = $this->workspaceService->getUserRoleInWorkspace($workspace, $currentUser->id);
 
         if ($currentUserRole !== 'owner') {
-            return response()->json(['message' => 'Доступ запрещен. Только владелец может добавлять пользователей'], 403);
+            return response()->json(['message' => __('messages.access_denied')], 403);
         }
 
-        $userId = $request->user_id;
-        $role = $request->input('role', 'executor');
+        $userId = $validate['user_id'];
+        $role = $validate['role'];
 
         if ($userId === $currentUser->id) {
-            return response()->json(['message' => 'Вы не можете изменить свою роль'], 403);
+            return response()->json(['message' => __('messages.cannot_change_own_role')], 403);
         }
 
         $this->workspaceService->manageUserInWorkspace($workspace, $userId, $role);
 
-        return response()->json(['message' => 'Успешно'], 201);
+        return response()->json(['message' => __('messages.success')], 201);
     }
 }
